Add benchmarking script

+++ src/fannkuch_rayon.rs
@@ -7,7 +7,7 @@ use rayon::prelude::*;
 // the number of permutations is lower than this value.
 const NUM_BLOCKS: u32 = 24;
 
-pub fn fannkuch_rayon(n: i32) -> (i32, i32) {
+pub fn fannkuch_rayon(n: usize) -> (i32, i32) {
     // Precompute a table a factorials to reuse all over the place.
     let mut factorials = [1; 16];
     for i in 1..n as usize + 1 {
